requirements.txt
databases
fastapi
feedparser
fyers_apiv3
requests
httpx
#numpy<1.24
numpy
openai
pandas
pandas_market_calendars
plotly
psycopg2
pydantic
pyotp
python-dotenv
pytz
scipy
SQLAlchemy
starlette
uvicorn
yfinance
jinja2
python-multipart
asyncpg
kiteconnect
indexed-gzip
